+++ Kloudy/kloudyWidget/Assets.xcassets/LockScreenWidgetImage/lockWarning.imageset/Contents.json
@@ -1,0 +1,12 @@
+{
+  "images" : [
+    {
+      "filename" : "lockWarning.png",
+      "idiom" : "universal"
+    }
+  ],
+  "info" : {
+    "author" : "xcode",
+    "version" : 1
+  }
+}
