+++ Kloudy/kloudyWidget/Assets.xcassets/LockScreenWidgetImage/lock_carwash_1.imageset/Contents.json
