+++ Kloudy/kloudyWidget/Assets.xcassets/LockScreenWidgetImage/lock_carwash_2.imageset/Contents.json
