+++ Kloudy/kloudyWidget/Assets.xcassets/LockScreenWidgetImage/lock_carwash_3.imageset/Contents.json
