+++ Kloudy/kloudyWidget/Assets.xcassets/LockScreenWidgetImage/lock_carwash_4.imageset/Contents.json
