+++ Kloudy/kloudyWidget/KloudyWarningWidget/KloudyWarningWidget.swift
@@ -35,3 +35,5 @@ struct KloudyWarningWidget: View {
     }
 }
 
+
+
